pyproject.toml
[project]
name = "pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "backoff==2.2.1",
    "boto3==1.26.122",
    "botocore==1.29.122",
    "celery==5.3.1",
    "click==8.1.3",
    "colorama>=0.4.6",
    "flask==2.3.1",
    "flask-cors==4.0.0",
    "flask-sqlalchemy==3.0.3",
    "invoke==2.2.0",
    "langchain==0.0.352",
    "langchain-community==0.0.19",
    "langchain-openai==0.0.5",
    "langfuse==1.0.18",
    "pinecone-client==3.2.2",
    "pypdf==3.15.4",
    "python-dotenv>=1.1.1",
    "redis==5.0.0",
    "tiktoken>=0.5.2",
    "uuid==1.30",
    "watchdog==3.0.0",
    "werkzeug==2.3.1",
]
